+++ src/main/java/net/earlystage/compat/SieveEmiRecipe.java
@@ -9,6 +9,7 @@ import dev.emi.emi.api.widget.WidgetHolder;
 import net.earlystage.EarlyStageMain;
 import net.earlystage.data.SieveDropTemplate;
 import net.minecraft.recipe.Ingredient;
+import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
 
 import java.util.ArrayList;
@@ -16,18 +17,19 @@ import java.util.List;
 
 public class SieveEmiRecipe implements EmiRecipe {
 
+    private final SieveDropTemplate template;
     private final List<EmiIngredient> inputs;
     private final List<EmiStack> outputs;
     private final Identifier id;
 
     public SieveEmiRecipe(SieveDropTemplate template) {
+        this.template = template;
         this.id = EarlyStageMain.identifierOf("/sieve/" + template.getBlockItem().toString().replace(':', '_'));
         this.inputs = List.of(EmiIngredient.of(Ingredient.ofItems(template.getBlockItem())));
 
         this.outputs = new ArrayList<>();
         for (int i = 0; i < template.getBlockDrops().size(); i++) {
             EmiStack stack = EmiStack.of(template.getBlockDrops().get(i));
-            stack.setChance(template.getDropChances().get(i));
             this.outputs.add(stack);
         }
     }
@@ -71,7 +73,7 @@ public class SieveEmiRecipe implements EmiRecipe {
         for (int i = 0; i < outputs.size(); i++) {
             int x = 48 + (i % 4) * 18;
             int y = 4 + (i / 4) * 18;
-            widgets.addSlot(outputs.get(i), x, y).recipeContext(this);
+            widgets.addSlot(outputs.get(i), x, y).recipeContext(this).appendTooltip(Text.translatable("emi.sieve.tooltip", Math.round(this.template.getDropChances().get(i) * 1000f) / 10f));
         }
     }
 }