+++ src/main/java/net/earlystage/item/WoodenBucketItem.java
@@ -1,14 +1,9 @@
 package net.earlystage.item;
 
+import net.earlystage.init.CompatInit;
 import net.earlystage.init.ItemInit;
 import net.minecraft.advancement.criterion.Criteria;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.Blocks;
-import net.minecraft.block.FluidBlock;
-import net.minecraft.block.FluidDrainable;
-import net.minecraft.block.FluidFillable;
-import net.minecraft.entity.Entity;
+import net.minecraft.block.*;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.fluid.FlowableFluid;
 import net.minecraft.fluid.Fluid;
@@ -18,6 +13,7 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ItemUsage;
 import net.minecraft.particle.ParticleTypes;
+import net.minecraft.registry.Registries;
 import net.minecraft.registry.tag.FluidTags;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.sound.SoundCategory;
@@ -25,6 +21,7 @@ import net.minecraft.sound.SoundEvent;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.stat.Stats;
 import net.minecraft.util.Hand;
+import net.minecraft.util.Identifier;
 import net.minecraft.util.TypedActionResult;
 import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.hit.HitResult;
@@ -61,7 +58,7 @@ public class WoodenBucketItem extends Item implements FluidModificationItem {
             }
             if (this.fluid == Fluids.EMPTY) {
                 BlockState blockState = world.getBlockState(blockPos);
-                if (blockState.getBlock() instanceof FluidDrainable && blockState.getFluidState().isOf(Fluids.WATER) && blockState.contains(FluidBlock.LEVEL)
+                if (blockState.getBlock() instanceof FluidDrainable drainable && blockState.getFluidState().isOf(Fluids.WATER) && blockState.contains(FluidBlock.LEVEL)
                         && blockState.get(FluidBlock.LEVEL) == 0) {
 
                     user.incrementStat(Stats.USED.getOrCreateStat(this));
@@ -69,10 +66,15 @@ public class WoodenBucketItem extends Item implements FluidModificationItem {
                     world.emitGameEvent(user, GameEvent.FLUID_PICKUP, blockPos);
 
                     ItemStack itemStack2 = new ItemStack(ItemInit.WATER_WOODEN_BUCKET, 1);
+                    if (CompatInit.PURIFIED_WATER_BRICK_BUCKET != null && blockState.getFluidState().isOf(Registries.FLUID.get(Identifier.of("dehydration:purified_water")))) {
+                        world.playSound(user, blockPos, SoundEvents.ITEM_BUCKET_FILL, SoundCategory.BLOCKS, 1.0f, 1.0f);
+                        itemStack2 = new ItemStack(CompatInit.PURIFIED_WATER_BRICK_BUCKET);
+                    }
                     ItemStack itemStack3 = ItemUsage.exchangeStack(itemStack, user, itemStack2);
 
                     if (!world.isClient()) {
-                        world.setBlockState(blockPos, Blocks.AIR.getDefaultState(), Block.NOTIFY_ALL | Block.REDRAW_ON_MAIN_THREAD);
+                        drainable.tryDrainFluid(user, world, blockPos, blockState);
+
                         Criteria.FILLED_BUCKET.trigger((ServerPlayerEntity) user, itemStack2);
                     }
                     return TypedActionResult.success(itemStack3, world.isClient());
@@ -131,7 +133,7 @@ public class WoodenBucketItem extends Item implements FluidModificationItem {
             return true;
         }
         if (block instanceof FluidFillable && this.fluid == Fluids.WATER) {
-            ((FluidFillable) ((Object) block)).tryFillWithFluid(world, pos, blockState, ((FlowableFluid) this.fluid).getStill(false));
+            ((FluidFillable) block).tryFillWithFluid(world, pos, blockState, ((FlowableFluid) this.fluid).getStill(false));
             this.playEmptyingSound(player, world, pos);
             return true;
         }
@@ -148,6 +150,6 @@ public class WoodenBucketItem extends Item implements FluidModificationItem {
     private void playEmptyingSound(@Nullable PlayerEntity player, WorldAccess world, BlockPos pos) {
         SoundEvent soundEvent = this.fluid.isIn(FluidTags.LAVA) ? SoundEvents.ITEM_BUCKET_EMPTY_LAVA : SoundEvents.ITEM_BUCKET_EMPTY;
         world.playSound(player, pos, soundEvent, SoundCategory.BLOCKS, 1.0f, 1.0f);
-        world.emitGameEvent((Entity) player, GameEvent.FLUID_PLACE, pos);
+        world.emitGameEvent(player, GameEvent.FLUID_PLACE, pos);
     }
 }